update tests and utilities
