+++ core/templates/pages/exploration-editor-page/modal-templates/exploration-save-prompt-modal.component.html
@@ -1,8 +1,8 @@
 <div class="modal-header e2e-test-save-prompt-modal">
-  <h4 class="modal-title">Save Changes</h4>
+  <h4 class="modal-title e2e-test-modal-header">Save Changes</h4>
 </div>
 <div class="modal-body">
-  <p>
+  <p class=".e2e-test-modal-body">
     It is recommended to save if the exploration has more than 50 changes.
   </p>
 </div>
