+++ core/templates/pages/exploration-editor-page/modal-templates/save-version-mismatch-modal.component.html
@@ -1,5 +1,5 @@
 <div class="modal-header e2e-test-lost-changes-modal">
-  <h3>Error Saving Exploration</h3>
+  <h3 class="e2e-test-modal-header">Error Saving Exploration</h3>
 </div>
 
 <div class="modal-body">
