+++ core/tests/puppeteer-acceptance-tests/utilities/user/exploration-editor.ts
@@ -442,7 +442,6 @@ const explorationGraphNodeSelector = 'g.e2e-test-node';
 const confirmDeleteStateButtonSelector = '.e2e-test-confirm-delete-state';
 
 const lostChangesModalSelector = '.e2e-test-lost-changes-modal';
-const lostChangesModalHeadingSelector = '.e2e-test-lost-changes-modal h3';
 const discardAndExportLostChangesButtonSelector =
   'button.e2e-test-discard-and-export-lost-changes-button';
 const stateNameSubmitButtonSelector = 'button.e2e-test-state-name-submit';
@@ -450,10 +449,10 @@ const stateNameInputSelector = '.e2e-test-state-name-input';
 const cardHeightLimitWarningSelector = '.e2e-test-card-height-limit-warning';
 
 const saveRecommendationModalSelector = '.e2e-test-save-prompt-modal';
-const saveRecommendationModalHeadingSelector = '.e2e-test-save-prompt-modal h4';
-const saveRecommendationModalBodyTextSelector = '.modal-body p';
 const saveRecommendationModalSaveButtonSelector =
   'button.e2e-test-recommendation-prompt-save-button';
+const profileDropdown = '.e2e-test-profile-dropdown';
+const creatorDashboardMenuLink = '.e2e-test-creator-dashboard-link';
 
 export enum INTERACTION_TYPES {
   ALGEBRAIC_EXPRESSION = 'Algebraic Expression Input',
@@ -7183,10 +7182,55 @@ export class ExplorationEditor extends BaseUser {
   }
 
   /**
+   * Clicks on a node element in the exploration graph
+   * @param stateName The name of the state to be clicked
+   * @param selectorToClick The selector of the element to be clicked within the node
+   * @returns A boolean indicating if the click was successful
+   */
+  async clickOnGraphNodeElement(
+    stateName: string,
+    selectorToClick: string
+  ): Promise<boolean> {
+    return await this.page.evaluate(
+      (name: string, selectorToClick: string, nodeSelector: string) => {
+        const nodes = Array.from(document.querySelectorAll(nodeSelector));
+
+        const targetNode = nodes.find(node => {
+          const title = node.querySelector('title')?.textContent || '';
+          const normalizedTitle = title.replace(/\s+/g, ' ').trim();
+          const normalizedName = name.replace(/\s+/g, ' ').trim();
+          return normalizedTitle.includes(normalizedName);
+        });
+
+        if (!targetNode) {
+          return false;
+        }
+
+        const element = targetNode.querySelector(
+          selectorToClick
+        ) as HTMLElement | null;
+
+        if (!element) {
+          return false;
+        }
+
+        element.scrollIntoView({block: 'center', inline: 'center'});
+        element.dispatchEvent(new Event('click', {bubbles: true}));
+        return true;
+      },
+      stateName,
+      selectorToClick,
+      explorationGraphNodeSelector
+    );
+  }
+
+  /**
    * Deletes a state from the exploration
    * @param stateName The name of the state to be deleted
    */
+
   async deleteState(stateName: string): Promise<void> {
+    await this.expectExplorationGraphToContainCard(stateName);
     const isMobile = this.isViewportAtMobileWidth();
 
     if (isMobile) {
@@ -7197,41 +7241,6 @@ export class ExplorationEditor extends BaseUser {
 
     await this.expectElementToBeVisible(explorationGraphSelector);
 
-    const clickOnGraphNodeElement = async (
-      selectorToClick: string
-    ): Promise<boolean> => {
-      return await this.page.evaluate(
-        (name: string, selectorToClick: string, nodeSelector: string) => {
-          const nodes = Array.from(document.querySelectorAll(nodeSelector));
-
-          const targetNode = nodes.find(node => {
-            const title = node.querySelector('title')?.textContent || '';
-            const normalizedTitle = title.replace(/\s+/g, ' ').trim();
-            const normalizedName = name.replace(/\s+/g, ' ').trim();
-            return normalizedTitle.includes(normalizedName);
-          });
-
-          if (!targetNode) {
-            return false;
-          }
-
-          const element = targetNode.querySelector(
-            selectorToClick
-          ) as HTMLElement | null;
-
-          if (!element) {
-            return false;
-          }
-          element.scrollIntoView({block: 'center', inline: 'center'});
-          element.dispatchEvent(new Event('click', {bubbles: true}));
-          return true;
-        },
-        stateName,
-        selectorToClick,
-        explorationGraphNodeSelector
-      );
-    };
-
     await this.page.waitForFunction(
       (nodeSelector: string) =>
         document.querySelectorAll(nodeSelector).length > 0,
@@ -7239,11 +7248,16 @@ export class ExplorationEditor extends BaseUser {
       explorationGraphNodeSelector
     );
 
-    await clickOnGraphNodeElement(explorationGraphNodeBackgroundSelector);
+    await this.clickOnGraphNodeElement(
+      stateName,
+      explorationGraphNodeBackgroundSelector
+    );
 
-    const deleteClicked = await clickOnGraphNodeElement(
+    const deleteClicked = await this.clickOnGraphNodeElement(
+      stateName,
       explorationGraphNodeDeleteButtonSelector
     );
+
     if (!deleteClicked) {
       throw new Error(
         `Delete button not found in graph for card: ${stateName}`
@@ -7284,15 +7298,12 @@ export class ExplorationEditor extends BaseUser {
     await this.expectElementToBeVisible(lostChangesModalSelector, isVisible);
 
     if (isVisible) {
-      await this.expectTextContentToContain(
-        lostChangesModalHeadingSelector,
-        'Error Saving Exploration'
-      );
+      await this.expectModalTitleToBe('Error Saving Exploration');
     }
   }
 
   /**
-   * Exports and discards lost changes and verifies download of lostChanges.txt
+   * Exports and discards lost changes from the lost changes modal
    */
   async exportAndDiscardLostChanges(): Promise<void> {
     await this.expectLostChangesModalToBeVisible(true);
@@ -7306,11 +7317,16 @@ export class ExplorationEditor extends BaseUser {
       discardAndExportLostChangesButtonSelector
     );
 
-    const downloadedFile =
-      await this.waitForExplorationDownload('lostChanges.txt');
-    expect(downloadedFile).toBe('lostChanges.txt');
-    await this.waitForPageToFullyLoad();
     await this.expectLostChangesModalToBeVisible(false);
+  }
+
+  /**
+   * Expects a file to be downloaded with given file name
+   * @param fileName the name of the file to be downloaded
+   */
+  async expectFileToBeDownloaded(fileName: string): Promise<void> {
+    const downloadedFile = await this.waitForExplorationDownload(fileName);
+    expect(downloadedFile).toBe(fileName);
   }
 
   /**
@@ -7318,6 +7334,7 @@ export class ExplorationEditor extends BaseUser {
    * @param newStateName - The new name for the state
    */
   async updateStateName(newStateName: string): Promise<void> {
+    await this.expectElementToBeVisible(currentCardNameContainerSelector, true);
     await this.clickOnElementWithSelector(currentCardNameContainerSelector);
 
     await this.page.evaluate(selector => {
@@ -7383,6 +7400,7 @@ export class ExplorationEditor extends BaseUser {
    * Clicks on the discard lost changes button
    */
   async discardLostChanges(): Promise<void> {
+    await this.expectElementToBeVisible(errorSavingExplorationModal);
     await this.clickOnElementWithSelector(errorSavingExplorationModal);
     await this.expectLostChangesModalToBeVisible(false);
   }
@@ -7403,13 +7421,9 @@ export class ExplorationEditor extends BaseUser {
   async expectSaveRecommendationModalToBeVisible(): Promise<void> {
     await this.expectElementToBeVisible(saveRecommendationModalSelector, true);
 
-    await this.expectTextContentToContain(
-      saveRecommendationModalHeadingSelector,
-      'Save Changes'
-    );
+    await this.expectModalTitleToBe('Save Changes');
 
-    await this.expectTextContentToContain(
-      saveRecommendationModalBodyTextSelector,
+    await this.expectModalContentToContain(
       'It is recommended to save if the exploration has more than 50 changes.'
     );
   }
@@ -7420,6 +7434,7 @@ export class ExplorationEditor extends BaseUser {
   async saveExplorationDraftFromSaveRecommendationModal(
     commitMessage: string = 'Testing Testing'
   ): Promise<void> {
+    await this.expectSaveRecommendationModalToBeVisible();
     await this.clickOnElementWithSelector(
       saveRecommendationModalSaveButtonSelector
     );
@@ -7437,6 +7452,18 @@ export class ExplorationEditor extends BaseUser {
     showMessage('Exploration is saved successfully.');
     await this.waitForPageToFullyLoad();
     await this.expectElementToBeVisible(saveRecommendationModalSelector, false);
+  }
+
+  /**
+   * Navigates to creator dashboard using profile dropdown.
+   */
+  async navigateToCreatorDashboardUsingProfileDropdown(): Promise<void> {
+    await this.expectElementToBeVisible(profileDropdown);
+    await this.clickOnElementWithSelector(profileDropdown);
+
+    await this.expectElementToBeVisible(creatorDashboardMenuLink);
+    await this.clickOnElementWithSelector(creatorDashboardMenuLink);
+    await this.expectElementToBeVisible(creatorDashboardContainerSelector);
   }
 }
